+++ packages/ripple-tide-api/src/nuxt/lib/utils.ts
@@ -1,9 +1,0 @@
-export const getBaseUrl = (req) => {
-  const apiPrefix = '/api/tide'
-  if (req) {
-    const protocol = req.protocol ? req.protocol : 'http'
-    return `${protocol}://${req.headers.host}${apiPrefix}`
-  } else if (typeof window !== 'undefined') {
-    return window.location.origin + apiPrefix
-  }
-}