+++ packages/ripple-tide-api/src/services/index.ts
@@ -1,3 +1,3 @@
 export { default as TideApiBase } from './tide-api-base.js'
-export { default as TidePage } from './tide-page.js'
-export { default as TideSite } from './tide-site.js'
+export { default as TidePageApi } from './tide-page.js'
+export { default as TideSiteApi } from './tide-site.js'