+++ packages/ripple-tide-event/components/TideEventHeader.vue
@@ -1,0 +1,17 @@
+<template>
+  <RplHeaderHeroHeader :corner-top="true" :corner-bottom="true" :behind-nav="true" :breadcrumbs="hasBreadcrumbs"
+    :title="header.title">
+    <p class="rpl-type-p-large">{{ header.summary }}</p>
+  </RplHeaderHeroHeader>
+</template>
+
+<script setup lang="ts">
+import type { TideEventHeader } from '../types'
+
+interface Props {
+  header: TideEventHeader
+  hasBreadcrumbs: boolean
+}
+
+defineProps<Props>()
+</script>